DEV: Improve logging for system commands
Instead of logging system commands for mc calls to STDOUT,
it's better to just log them to the same place as the server
log, since they are mostly related to the minio server. This
will stop messages like this coming up in rspec output:

Added `local` successfully.
[2023-08-15 15:57:54 AEST]     0B optimized/
[2023-08-15 15:57:54 AEST]     0B original/
[2023-08-15 15:57:54 AEST]     0B temp/
Access permission for `local/discoursetest` is set to `public`
      uploads a file in the post composer